requirements.txt
Django==1.8.6
django-debug-toolbar==1.4
django-paypal==0.2.7
django-registration-redux==1.2
django-sslserver==0.16
djrill==1.4.0
requests==2.8.1
six==1.10.0
sqlparse==0.1.18
stripe==1.27.1
wheel==0.24.0
